ENH: Initial Commit for BRAINSDeface
Adds a readme file with a brief description as the initial commit to the
BRAINSDeface project.

ENH: Adds basic cmake infrastructure for new module BRAINSDeface

BRAINSDeface was added to the list of modules in BRAINSTools.cmake,
and the USE_BRAINSDeface option was added to Common.cmake. A simple
CMakeLists.txt file was added to the BRAINSDeface directory

ENH: Adds "Hello World" program and basic xml file to BRAINSDeface

A simple "Hello World" file and an SEM xml file were added to the
BRAINSDeface module. The CMakeLists file was updated.