ENH: Cleanup, refactoring, for module BRAINSDeface
ENH: Adds Write Transform convienience function

Refactors writing of transforms to disk into a utility function so that is can
be done with one line of code.

DOC: Removes unused and old commented out code

ENH: Provides a model for an ITK style filter

We anticipate that ITK will be heavily used in our project. Some of our
classes may need to be a part of the ITK pipline. This basic example provides
a model that may be used when writing classes that conform to the ITK standard
and use Smart Pointers ->SetInput ->GetOutput ->Update functionality etc. and
can be used within the ITK pipeline.