ENH: Adds initial Test Suite and unit tests for BRAINSDeface module!
ENH: Adds initial unit test

Added unit test based on BRAINSTools style for CTest

ENH: Adds a basic "proof of runnability" test

A basic "does it run" test is added to the BRAINSDeface module. It
checks to see if the program can be run. Our current testing data is not public
and so cannot be hosted on midas. Instead it is being put into:
/scratch/testData/BRAINSDefaceData
We should add a cmake argument that can change this directory so its specified
at configure time.

ENH: Adds unit test to check if brain data has been altered

Uses the brain label map to mask the brain of the original image and defaced image.
Then it uses ITK absolute value difference image filter to create a difference image.
Then all the values of the stistics image is summed together, and the tests fails if the sum is not 0.